marked zoomLevel as double

+++ src/main/java/com/chiragji/docuconverter/logic/PptToPdf.java
@@ -21,9 +21,9 @@ import java.util.List;
 public final class PptToPdf extends AbstractDocuConverter {
     private List<HSLFSlide> slides;
     private Dimension size;
-    private int zoomLevel = 2;
+    private double zoomLevel = 2;
 
-    public void setZoomLevel(int zoomLevel) {
+    public void setZoomLevel(double zoomLevel) {
         if (zoomLevel <= 0)
             throw new IllegalArgumentException("Invalid zoo level: " + zoomLevel +
                     ", A non zero positive number is only acceptable");
